DEV: move AI to use zeitwerk instead of custom loading
