+++ lib/discourse_ai/ai_bot/anthropic_bot.rb
