+++ lib/discourse_ai/ai_bot/bot.rb
