+++ lib/discourse_ai/ai_bot/commands/categories_command.rb
