+++ lib/discourse_ai/ai_bot/commands/command.rb
