+++ lib/discourse_ai/ai_bot/commands/dall_e_command.rb
