+++ lib/discourse_ai/ai_bot/commands/db_schema_command.rb
