+++ lib/discourse_ai/ai_bot/commands/google_command.rb
