+++ lib/discourse_ai/ai_bot/commands/image_command.rb
