+++ lib/discourse_ai/ai_bot/commands/read_command.rb
