+++ lib/discourse_ai/ai_bot/commands/search_command.rb
