+++ lib/discourse_ai/ai_bot/commands/search_settings_command.rb
