+++ lib/discourse_ai/ai_bot/commands/setting_context_command.rb
