+++ lib/discourse_ai/ai_bot/commands/summarize_command.rb
