+++ lib/discourse_ai/ai_bot/commands/tags_command.rb
