+++ lib/discourse_ai/ai_bot/commands/time_command.rb
