+++ lib/discourse_ai/ai_bot/open_ai_bot.rb
