+++ lib/discourse_ai/ai_bot/personas/artist.rb
