+++ lib/discourse_ai/ai_bot/personas/creative.rb
