+++ lib/discourse_ai/ai_bot/personas/dall_e_3.rb
