+++ lib/discourse_ai/ai_bot/personas/general.rb
