+++ lib/discourse_ai/ai_bot/personas/researcher.rb
