+++ lib/discourse_ai/ai_bot/personas/settings_explorer.rb
