+++ lib/discourse_ai/ai_bot/personas/sql_helper.rb
