+++ lib/discourse_ai/ai_bot/site_settings_extension.rb
