+++ lib/discourse_ai/engine.rb
@@ -4,5 +4,6 @@ module ::DiscourseAi
   class Engine < ::Rails::Engine
     engine_name PLUGIN_NAME
     isolate_namespace DiscourseAi
+    config.autoload_paths << File.join(config.root, "lib")
   end
 end
